xpayback1/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
icecream==2.1.3
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
